+++ crates/acts/src/model/branch.rs
@@ -95,11 +95,11 @@ impl Branch {
         self.options
             .entry(consts::ACT_EXPOSE)
             .and_modify(|outputs| {
-                if let Some(obj) = outputs.as_object_mut() {
-                    obj.insert(name.to_string(), json!(value));
+                if let Some(obj) = outputs.as_array_mut() {
+                    obj.push(json!(Variant::create(name, value.clone())));
                 }
             })
-            .or_insert(Vars::new().with(name, value).into());
+            .or_insert(json!(vec![Variant::create(name, value)]));
 
         self
     }