+++ crates/acts/src/model/tests/branch.rs
@@ -1,4 +1,4 @@
-use crate::{Branch, Vars, Workflow, utils::consts};
+use crate::{Branch, Workflow};
 use serde_json::json;
 
 #[test]
@@ -24,8 +24,7 @@ fn model_branch_vars() {
 fn model_branch_outputs() {
     let b = Branch::new().with_expose("p1", json!(5));
 
-    let options = b.options.get::<Vars>(consts::ACT_EXPOSE).unwrap();
-    assert_eq!(options.len(), 1);
+    let options = b.exposes();
     assert!(options.get_value("p1").is_some());
 }
 
@@ -127,7 +126,7 @@ fn model_branch_yml_expose() {
             - id: b1
               if: true
               options:
-                expose:
+                exposes:
                     - name: p1
 
     "#;
@@ -138,4 +137,10 @@ fn model_branch_yml_expose() {
     let exposes = barnch.exposes();
     assert_eq!(exposes.len(), 1);
     assert_eq!(exposes.get_value("p1"), Some(&json!(null)));
+}
+
+#[test]
+fn model_branch_with_expose() {
+    let b = Branch::new().with_expose("v1", 0);
+    assert!(b.exposes().contains_key("v1"));
 }