+++ crates/acts/src/model/workflow.rs
@@ -1,5 +1,6 @@
 use crate::{
-    Act, ActError, ActValue, ModelBase, Result, Step, Variant, Vars, scheduler::NodeTree, utils::consts,
+    Act, ActError, ActSchema, ModelBase, Result, Step, Variant, Vars, scheduler::NodeTree,
+    utils::consts,
 };
 use serde::{Deserialize, Serialize};
 use serde_json::json;
@@ -30,11 +31,11 @@ pub struct Workflow {
 
     /// input json schema
     #[serde(default)]
-    pub inputs: ActValue,
+    pub inputs: ActSchema,
 
     /// output json schema
     #[serde(default)]
-    pub outputs: ActValue,
+    pub outputs: ActSchema,
 
     #[serde(default)]
     pub on: Vec<Act>,
@@ -210,12 +211,12 @@ impl Workflow {
         self
     }
 
-    pub fn with_inputs(mut self, inputs: ActValue) -> Self {
+    pub fn with_inputs(mut self, inputs: ActSchema) -> Self {
         self.inputs = inputs;
         self
     }
 
-    pub fn with_outputs(mut self, outputs: ActValue) -> Self {
+    pub fn with_outputs(mut self, outputs: ActSchema) -> Self {
         self.outputs = outputs;
         self
     }
@@ -227,19 +228,19 @@ impl Workflow {
         self.options
             .entry(consts::ACT_EXPOSE)
             .and_modify(|outputs| {
-                if let Some(obj) = outputs.as_object_mut() {
-                    obj.insert(name.to_string(), json!(value));
+                if let Some(obj) = outputs.as_array_mut() {
+                    obj.push(json!(Variant::create(name, value.clone())));
                 }
             })
-            .or_insert(Vars::new().with(name, value).into());
+            .or_insert(json!(vec![Variant::create(name, value)]));
 
         self
     }
 
     pub fn exposes(&self) -> Vars {
         let mut vars = Vars::new();
-        if let Some(expose) = self.options.get::<Vec<Variant>>(consts::ACT_EXPOSE) {
-            expose.iter().for_each(|var| {
+        if let Some(exposes) = self.options.get::<Vec<Variant>>(consts::ACT_EXPOSE) {
+            exposes.iter().for_each(|var| {
                 vars.set(&var.name, var.value.clone());
             });
         }